backend/requirements.txt
fastapi
uvicorn
pydicom
pillow
requests
openai      # optional if you use GPT
